Move Relations adapter and rename it (Adapter)
